requirements.txt
# Core dependencies
numpy>=1.20.0
opencv-python>=4.5.0
ultralytics>=8.0.0  # For YOLOv8
face-recognition>=1.3.0
pymongo>=4.0.0
psycopg2-binary>=2.9.0
cryptography>=37.0.0

# Web framework
flask>=2.0.0
flask-cors>=3.0.10

# Utilities
python-dotenv>=0.20.0
pyyaml>=6.0
tqdm>=4.64.0
pillow>=9.0.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0